+++ packages/ui/CHANGELOG.md
@@ -6,7 +6,7 @@ The following is a curated list of changes in the Enact ui module, newest change
 
 ### Added
 
-- `ui/Marquee` prop `marqueePadding` and CSS class `padding` to configure the padding between the repeated content
+- `ui/Marquee` prop `marqueeSpacing` and CSS class `spacing` to configure the spacing between the repeated content
 
 ## [3.1.2] - 2019-09-30